# doc/swagger-api-description/api.yaml
swagger: '2.0'
info:
  title: ariADDna API
  description: |
   #### This document contains the API description for ariADDna project.
   Using this API one can manage all available cloud services (DropBox, GDrive, Yandex.Disk etc.) from single point.
  contact:
    name: API Support
    url: http://www.ariaddna.stnetix.com/support
    email: ariaddna.support@stnetix.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
host: api.ariaddna.stnetix.com
basePath: /v1
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /vufs/snap/{userUuid}:
    get:
      description: Allows to get snapshot vufs.
      operationId: getVUFS
      parameters:
        - in: path
          name: userUuid
          description: Current user UUID.
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: VUFS snapshot of current user.
          schema:
            $ref: '#/definitions/Vufs'
        '404':
          description: Bad user UUID.
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/ErrorModel'
  /vufs/snap/diff/{userUuid}/{dateTime}:
    get:
      description: Allows to get difference of previous snapshot and actual.
      operationId: getDiffVUFS
      parameters:
        - in: path
          name: userUuid
          description: Current user UUID.
          required: true
          type: string
          format: uuid
        - in: path
          name: dateTime
          description: FromDateTime of current VUFS snapshot
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Difference VUFS snapshot of current user.
          schema:
            $ref: '#/definitions/Vufs'
        '404':
          description: Bad user UUID.
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/ErrorModel'
  /vufs/snap/diff/{userUuid}:
    post:
      description: Allows to send changes in local file storage to server with empty Allocation model and as response get Vufs object with Allocation model.
      operationId: sendChangesInLFS
      parameters:
        - in: body
          name: localChanges
          description: VUFS snapshot typed object with changed in local file storage with empty Allocation model.
          required: true
          schema:
           $ref: '#/definitions/Vufs'
        - in: path
          name: userUuid
          description: Current user UUID.
          required: true
          type: string
          format: uuid
      responses:
        '201':
          description: Checked snapshot changes.
          schema:
           $ref: '#/definitions/Vufs'
        '409':
          description: Changes was not found in clouds.
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/ErrorModel'
  /vufs/allocation/{userUuid}:
    post:
      description: Allows to post file allocate strategy from client to server.
      operationId: postAllocateModel
      parameters:
        - in: path
          name: userUuid
          description: User UUID.
          required: true
          type: string
          format: uuid
        - in: body
          name: initialAllocationModel
          description: File allocation strategy.
          required: true
          schema:
           $ref: '#/definitions/InitialAllocationModel'
      responses:
        '201':
          description: Allocation strategy was send to server. As respose - difference VUFS object.
          schema:
           $ref: '#/definitions/Vufs'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/ErrorModel'
...
definitions:
...
  AriaddnaObj:
    type: object
    description: Describes abstract arriadna object.
    properties:
      uuid:
        description: Object UUID.
        type: string
        format: uuid
    required:
      - uuid
  User:
    type: object
    description: User of service.
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      userCredential:
        description: User's credential.
        $ref: '#/definitions/Credential'
      nickname:
        type: string
        description: User's nickname.
      email:
        type: string
        description: User's e-mail.
      externalClouds:
        description: The set of external cloud accounts that user has at ariADDna server.
        type: object
        $ref: '#/definitions/CloudSet'
    required:
      - userCredential
      - nickname
      - externalClouds
  Credential:
    type: object
    description: Credential for a basic user/password authentication.
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      user:
        type: string
        description: User's login.
      password:
        type: string
        description: A password associated with the login.
        format: password
    required:
      - user
      - password
...
  Cloud:
    description: A representation of an external cloud object.
    type: object
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      cloudUrl:
        description: A base URL of an external cloud for file managment.
        type: string
      cloudType:
        description: Enumeration of all cloud services with which ariADDna can interact.
        enum:
          - GOOGLE_DRIVE
          - YANDEX_DISK
          - DROPBOX
      cloudAccessToken:
        description: The token which allows ariADDna service action at an external cloud service on behalf of a user.
        $ref: '#/definitions/CloudAccessToken'
      connectionState:
        description: Enumeration of certain connection states to an external cloud service.
        enum:
          - CONNECTED
          - DISCONNECTED
          - IN_PROGRESS
          - TERMINATED
      login:
        description: A login name for an external cloud service where a user is already registered.
        type: string
      password:
        description: A password for access to be granted.
        type: string
      availableSpace:
        description: Cloud available spase in bytes.
        type: integer
        format: int64
    required:
      - cloudType
      - cloudUrl
      - cloudAccessToken
      - availableSpace
  CloudSet:
    description: A representation of a cloud list that user has.
    type: array
    items:
      $ref: '#/definitions/Cloud'
...
  CloudAccessToken:
    description: Stands for an external cloud access token. For more info see https://tools.ietf.org/html/draft-ietf-oauth-v2-31.
    type: object
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      accessToken:
        description: A token recieved during OAuth authorization.
        type: string
      expiresIn:
        description: The life-time in seconds of the access token.
        type: integer
      refreshToken:
        description: If an accessToken has a limited life-time a user will need to pass authorization manualy again. To prevent this a refreshToken should be returned togeather with an accessToken. Thus having a refresh token and passing it inside HTTP-request ariADDna service will be able to get new accessToken without user participation.
        type: string
      tokenType:
        description: OPTIONAL. The access token type provides the client with the information required to successfully utilize the access token to make a protected resource request (along with type-specific attributes).  The client MUST NOT use an access token if it does not understand the token type. For example Yandex.Disk whats the tokenType to be 'bearer'.
        type: string
    required:
      - accessToken
  Vufs:
    description: Describes shapshot VUFS of user.
    type: object
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      fromDateTime:
        description: Created date of snapshot in milliseconds.
        type: integer
        format: int64
      toDateTime:
        description: Deactivation date of snapshot in milliseconds.
        type: integer
        format: int64
      owner:
        description: Owner of this snapshot.
        $ref: '#/definitions/User'
      vufsFileSet:
        type: array
        items:
          $ref: '#/definitions/VufsFile'
    required:
      - fromDateTime
      - owner
  VufsFile:
    description: Describes virtual file in VUFS.
    type: object
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      file-properties:
        type: array
        items:
          $ref: '#/definitions/Property'
      file-parts:
        type: array
        items:
          $ref: '#/definitions/PartLocation'
      childSet:
        type: array
        items:
          $ref: '#/definitions/VufsFile'
      allocateModel:
        $ref: '#/definitions/AllocationModel'
    required:
      - file-properties
      - file-parts
  Property:
    description: Describes all property.
    type: object
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      property-name:
        type: string
        description: Property name.
      property-value:
        type: string
        description: Property value.
    required:
      - property-name
      - property-value
  PartLocation:
    description: Describes sequence number of file part and its location.
    type: object
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      number:
        type: integer
        format: int32
        description: Sequence number.
      location:
        $ref: '#/definitions/Cloud'
    required:
      - number
      - location
  AllocationModel:
    description: Describes allocate storage model ot file.
    type: object
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      modelType:
        description: Types of allocation model.
        enum:
          - UNION
          - MIRROR
          - HIGH_AVAILABILITY
      propertySet:
        type: array
        items:
          $ref: '#/definitions/Property'
    required:
      - modelType
  InitialAllocationModel:
    type: object
    description: Contain information about initial file allocation stategy.
    allOf:
      - $ref: '#/definitions/AriaddnaObj'
    properties:
      file:
        $ref: '#/definitions/VufsFile'
      allocationModel:
        $ref: '#/definitions/AllocationModel'
    required:
      - file
      - allocationModel
...
  ErrorModel:
    description: Information about error.
    type: object
    properties:
      code:
        description: Contains error code massage.
        type: integer
        format: int32
      message:
        description: Contains short error message.
        type: string
    required:
      - code
      - message

#Below all the items are just stubs.